add script for up5k 'wishbone-tool -s terminal...'
This automates the timing sensitive task of reseting the UP5K and
starting a wishbone-tool crossover UART bridge. Currently, the
connection needs to start in the time window between when the
bitstream loads (~50ms) and when the firmware locks up without a
bridge connection (~2000ms). The lockup may be due to the crossover
UART TX buffer filling up -- not sure.

Signed-off-by: samblenny