add more debug output to bbram_helper, re-introduce jtag delay
programming is not reliable against a build that has debug enabled
on the counter side without the delay. Negligle performance impact
on the key burning process, since the amount of data transferred
is so small.

Also added a dummy carriage return at the start of the bbram helper
in case the user had left the dialog box open instructing them to
start the program.